requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
transformers==4.49.0
accelerate==1.10.1
peft==0.17.1
scipy==1.15.3
torchdiffeq==0.2.5
qwen_vl_utils==0.0.11
PyYAML>=6.0
websockets>=12.0
msgpack>=1.0.5
msgpack-numpy>=0.4.8
